refactor(test): move billing helpers to db-test-seeders and db-test-assertions
Move billing/Stripe helpers from api-test-helpers/credits.ts into
categorized directories:

- db-test-seeders/credits.ts: 5 DB-direct seeders with @why-db-direct
  annotations (updateOrgStripeFields, updateOrgAutoRecharge,
  updateOrgStripeSubscription, insertTestCreditPricing,
  insertCreditExpiresRecord)
- db-test-assertions/credits.ts: 3 read-only assertion helpers
  (getOrgBillingFields, getOrgAutoRechargeFields, findCreditExpiresRecords)

api-test-helpers/credits.ts now contains only service-layer wrappers
(grantCreditsToOrg, testDeductFromExpiresRecords, testExpireCredits),
usage/insights helpers (Part 2), and re-exports for backward
compatibility.

Closes #9341
