+++ turbo/apps/web/src/__tests__/db-test-assertions/credits.ts
@@ -1,0 +1,60 @@
+import { eq } from "drizzle-orm";
+import { initServices } from "../../lib/init-services";
+import { orgMetadata } from "../../db/schema/org-metadata";
+import { creditExpiresRecord } from "../../db/schema/credit-expires-record";
+
+// ---------------------------------------------------------------------------
+// Read-only assertion helpers for billing / credit test verification.
+// ---------------------------------------------------------------------------
+
+/**
+ * Read all billing-related fields from an org in the `org_metadata` table.
+ */
+export async function getOrgBillingFields(orgId: string) {
+  initServices();
+  const [row] = await globalThis.services.db
+    .select({
+      tier: orgMetadata.tier,
+      credits: orgMetadata.credits,
+      stripeCustomerId: orgMetadata.stripeCustomerId,
+      stripeSubscriptionId: orgMetadata.stripeSubscriptionId,
+      subscriptionStatus: orgMetadata.subscriptionStatus,
+      currentPeriodEnd: orgMetadata.currentPeriodEnd,
+      cancelAtPeriodEnd: orgMetadata.cancelAtPeriodEnd,
+      lastProcessedInvoiceId: orgMetadata.lastProcessedInvoiceId,
+    })
+    .from(orgMetadata)
+    .where(eq(orgMetadata.orgId, orgId))
+    .limit(1);
+  return row ?? null;
+}
+
+/**
+ * Read auto-recharge fields from an org.
+ */
+export async function getOrgAutoRechargeFields(orgId: string) {
+  initServices();
+  const [row] = await globalThis.services.db
+    .select({
+      autoRechargeEnabled: orgMetadata.autoRechargeEnabled,
+      autoRechargeThreshold: orgMetadata.autoRechargeThreshold,
+      autoRechargeAmount: orgMetadata.autoRechargeAmount,
+      autoRechargePendingAt: orgMetadata.autoRechargePendingAt,
+    })
+    .from(orgMetadata)
+    .where(eq(orgMetadata.orgId, orgId))
+    .limit(1);
+  return row ?? null;
+}
+
+/**
+ * Find all credit expires records for an org, ordered by expires_at ASC.
+ */
+export async function findCreditExpiresRecords(orgId: string) {
+  initServices();
+  return globalThis.services.db
+    .select()
+    .from(creditExpiresRecord)
+    .where(eq(creditExpiresRecord.orgId, orgId))
+    .orderBy(creditExpiresRecord.expiresAt);
+}
