+++ turbo/apps/web/src/__tests__/db-test-seeders/credits.ts
@@ -1,0 +1,163 @@
+import { eq } from "drizzle-orm";
+import { initServices } from "../../lib/init-services";
+import { orgMetadata } from "../../db/schema/org-metadata";
+import { creditPricing } from "../../db/schema/credit-pricing";
+import { creditExpiresRecord } from "../../db/schema/credit-expires-record";
+
+// ---------------------------------------------------------------------------
+// DB-direct seeders for billing / Stripe test setup.
+//
+// Each function has a @why-db-direct annotation explaining why it cannot be
+// replaced by a webhook simulation or API call.
+// ---------------------------------------------------------------------------
+
+/**
+ * Set Stripe billing fields on an org in the `org_metadata` table.
+ *
+ * @why-db-direct Sets Stripe billing preconditions (stripeCustomerId,
+ * subscriptionId, tier). These fields are normally written by Stripe
+ * Dashboard / checkout flow. No API or webhook in our codebase bootstraps
+ * these from scratch — `handleCheckoutCompleted` READS `stripeCustomerId`
+ * to find the org, so it cannot create the initial association.
+ */
+export async function updateOrgStripeFields(
+  orgId: string,
+  fields: {
+    stripeCustomerId?: string | null;
+    stripeSubscriptionId?: string | null;
+    subscriptionStatus?: string | null;
+    currentPeriodEnd?: Date | null;
+    cancelAtPeriodEnd?: boolean;
+    lastProcessedInvoiceId?: string | null;
+    tier?: string;
+  },
+): Promise<void> {
+  initServices();
+  await globalThis.services.db
+    .update(orgMetadata)
+    .set({ ...fields, updatedAt: new Date() })
+    .where(eq(orgMetadata.orgId, orgId));
+}
+
+/**
+ * Configure auto-recharge settings on an org.
+ *
+ * @why-db-direct Auto-recharge configuration is normally set via the billing
+ * settings API which requires an active paid subscription. Direct seeding
+ * avoids multi-step subscription ceremony for test setup.
+ */
+export async function updateOrgAutoRecharge(
+  orgId: string,
+  fields: {
+    autoRechargeEnabled?: boolean;
+    autoRechargeThreshold?: number | null;
+    autoRechargeAmount?: number | null;
+    autoRechargePendingAt?: Date | null;
+  },
+): Promise<void> {
+  initServices();
+  await globalThis.services.db
+    .update(orgMetadata)
+    .set({ ...fields, updatedAt: new Date() })
+    .where(eq(orgMetadata.orgId, orgId));
+}
+
+/**
+ * Set Stripe subscription fields on org_metadata for testing billing-related flows.
+ *
+ * @why-db-direct Sets subscription ID and status for cleanup / deletion tests.
+ * A subset of updateOrgStripeFields. No webhook creates subscription
+ * associations from scratch.
+ */
+export async function updateOrgStripeSubscription(
+  orgId: string,
+  subscriptionId: string,
+  status: string,
+): Promise<void> {
+  initServices();
+  await globalThis.services.db
+    .update(orgMetadata)
+    .set({
+      stripeSubscriptionId: subscriptionId,
+      subscriptionStatus: status,
+      updatedAt: new Date(),
+    })
+    .where(eq(orgMetadata.orgId, orgId));
+}
+
+/**
+ * Insert a credit_pricing record for testing.
+ * Uses upsert so tests can safely set pricing for the same model.
+ *
+ * @why-db-direct Credit pricing is reference data managed via database
+ * migrations, not API endpoints or webhooks. No user-facing flow creates
+ * pricing records.
+ */
+export async function insertTestCreditPricing(
+  model: string,
+  options?: {
+    inputTokenPrice?: number;
+    outputTokenPrice?: number;
+    cacheReadTokenPrice?: number;
+    cacheCreationTokenPrice?: number;
+    modelProvider?: string;
+  },
+): Promise<void> {
+  initServices();
+  const inputTokenPrice = options?.inputTokenPrice ?? 100;
+  const outputTokenPrice = options?.outputTokenPrice ?? 200;
+  const cacheReadTokenPrice = options?.cacheReadTokenPrice ?? 0;
+  const cacheCreationTokenPrice = options?.cacheCreationTokenPrice ?? 0;
+  const modelProvider = options?.modelProvider ?? "";
+
+  await globalThis.services.db
+    .insert(creditPricing)
+    .values({
+      model,
+      modelProvider,
+      inputTokenPrice,
+      outputTokenPrice,
+      cacheReadTokenPrice,
+      cacheCreationTokenPrice,
+    })
+    .onConflictDoUpdate({
+      target: [creditPricing.model, creditPricing.modelProvider],
+      set: {
+        inputTokenPrice,
+        outputTokenPrice,
+        cacheReadTokenPrice,
+        cacheCreationTokenPrice,
+      },
+    });
+}
+
+/**
+ * Insert a credit expires record for testing.
+ *
+ * @why-db-direct Credit expires records are normally created by
+ * `handleInvoicePaid` during subscription renewal. Tests need precise
+ * control over amounts, remaining balances, and expiry dates that cannot
+ * be achieved through the webhook flow.
+ */
+export async function insertCreditExpiresRecord(params: {
+  orgId: string;
+  source?: string;
+  stripeInvoiceId?: string;
+  amount: number;
+  remaining?: number;
+  expiresAt: Date;
+}): Promise<string> {
+  initServices();
+  const [row] = await globalThis.services.db
+    .insert(creditExpiresRecord)
+    .values({
+      orgId: params.orgId,
+      source: params.source ?? "subscription_renewal",
+      stripeInvoiceId: params.stripeInvoiceId ?? null,
+      amount: params.amount,
+      remaining: params.remaining ?? params.amount,
+      expiresAt: params.expiresAt,
+    })
+    .returning({ id: creditExpiresRecord.id });
+  return row!.id;
+}
